Update class properties, refactor
Updated class properties on User and Company to be homogenous. Removed
second repeated method in CustomMap and instead created an interface to
check the type of incoming MapEntity's. So long as the provided data ca
n provide a location object with props lat and lng (what was changed on
the updated class props, just removed an o from long) it can be mapped.
Refreshed index.html on browser and it correctly generates two markers
as expected.